Improved logging and `util` code. Moved time-related utility functions to `utils.time`. Changed to module-name loggers in `updates`, `reminders` and `converters`. Edited logging messages in `reminders`. Fixed bug in `&reminder delete`.
